Backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.9.1
supabase==2.17.0
reportlab==4.4.3
pillow==11.3.0
httpx==0.28.1
psutil==6.1.0
PyJWT==2.10.1
bcrypt==4.1.2
requests==2.31.0